Jenkins-fix

+++ src/qt/widgets/properties/transferfunctionpropertywidgetqt.cpp
@@ -70,7 +70,7 @@ namespace inviwo {
 
     void TransferFunctionPropertyWidgetQt::setPropertyValue() {
         QColor color = colorDialog_->currentColor();
-        for (int i = 0; i < points_.size() ; i++){
+        for (int i = 0; i < (int)points_.size() ; i++){
             if (points_[i]->isSelected()){
                 points_[i]->getPoint()->setRgb(new vec3(color.redF(),color.greenF(),color.blueF()));
             }
@@ -82,7 +82,7 @@ namespace inviwo {
     }
 
 
-    void TransferFunctionPropertyWidgetQt::updateFromProperty() {
+    void TransferFunctionPropertyWidgetQt::updateFromProperty(){
         QPen pen;
         if (checkBox_->isChecked()){
             for (int i = 0; i < 256 ; i++){
